Add a new t-digest impl: BatchDigest.
This implementation is different from the ones we already have in the sense that
it does not try to always keep a sorted histogram. Instead it maintains a buffer
and a main histogram and periodically merges the buffer into the histogram.

Tests pass but I think it's still required to check this impl for correctness
given how different it is from the other impls.

Although it would not be a good fit if you need to evalutate quantiles
periodically (every time that you do so, it needs to merge the buffer into the
main histogram), it seems to perform much faster for the 'batch' case when you
first add all your values and then evaluate quantiles.