Prevent Centroid ID collisions
In a distributed system (or other environment consisting of multiple JVM instances) it is possible
for the current Centroid code to generate duplicate IDs. This breaks compareTo because the ID is
used as a tie-breaker. This commit changes the ID from an `int` to a `long`, and uses the high bits
to store a "JVM ID". This should make collisions statistically infeasible (one in 4 billion for a
given pair of JVMs).