Forced stable sorting. Closes #153, #114, #113, #102
Made the sort for MergingDigest stable. This avoids a
host of issues with repeated values that can violate
the digest invariants.

Also added some sanity tests into the digest merge
which required some tests to be made more sophisticated
so they don't create test digests that violate the
invariants.

This also required a number of tests that depended on
invalid digests to be adjusted.

This still carries along the old sorting code and the
new sort is insufficiently tested.